Editing problems

+++ src/main/kotlin/lt/skafis/bankas/service/implementations/ApprovalServiceImpl.kt
@@ -99,11 +99,15 @@ class ApprovalServiceImpl: ApprovalService {
         )
         sourceRepository.update(modifiedSource, sourceId)
 
-        val sourceProblems = problemRepository.getBySourceId(sourceId)
-        sourceProblems.forEach {
-            problemRepository.update(it.copy(skfCode = "", isApproved = false), it.id)
+        if (source.reviewStatus == ReviewStatus.APPROVED) {
+            val sourceProblems = problemRepository.getBySourceId(sourceId)
+            sourceProblems.forEach {
+                problemRepository.update(it.copy(skfCode = "", isApproved = false), it.id)
+                if (it.isApproved) {
+                    metaService.removeSkfCodeFromUsedList(it.skfCode)
+                }
+            }
         }
-
         return createdProblem.id
     }
 
@@ -153,16 +157,19 @@ class ApprovalServiceImpl: ApprovalService {
         )
         sourceRepository.update(updatedSource, sourceId)
 
-        val problems = problemRepository.getBySourceId(sourceId)
-        problems.forEach {
-            val skfCode = metaService.getLowestUnusedSkfCode()
-            metaService.amendUsedSkfCodeList(skfCode)
-            val updatedProblem = it.copy(
-                skfCode = skfCode,
-                isApproved = true
-            )
-            problemRepository.update(updatedProblem, it.id)
+        if (source.reviewStatus != ReviewStatus.APPROVED) {
+            val problems = problemRepository.getBySourceId(sourceId)
+            problems.forEach {
+                val skfCode = metaService.getLowestUnusedSkfCode()
+                metaService.amendUsedSkfCodeList(skfCode)
+                val updatedProblem = it.copy(
+                    skfCode = skfCode,
+                    isApproved = true
+                )
+                problemRepository.update(updatedProblem, it.id)
+            }
         }
+
         return updatedSource.toDisplayDto(userService.getUsernameById(updatedSource.authorId))
     }
 
@@ -175,14 +182,16 @@ class ApprovalServiceImpl: ApprovalService {
         )
         sourceRepository.update(updatedSource, sourceId)
 
-        val problems = problemRepository.getBySourceId(sourceId)
-        problems.forEach {
-            val updatedProblem = it.copy(
-                skfCode = "",
-                isApproved = false
-            )
-            problemRepository.update(updatedProblem, it.id)
-            metaService.removeSkfCodeFromUsedList(it.skfCode)
+        if (source.reviewStatus == ReviewStatus.APPROVED) {
+            val problems = problemRepository.getBySourceId(sourceId)
+            problems.forEach {
+                val updatedProblem = it.copy(
+                    skfCode = "",
+                    isApproved = false
+                )
+                problemRepository.update(updatedProblem, it.id)
+                metaService.removeSkfCodeFromUsedList(it.skfCode)
+            }
         }
         return updatedSource.toDisplayDto(userService.getUsernameById(updatedSource.authorId))
     }
@@ -214,18 +223,7 @@ class ApprovalServiceImpl: ApprovalService {
         if (problem.answerImagePath.startsWith("answers/")) {
             storageRepository.deleteImage(problem.answerImagePath)
         }
-        val modifiedSource = source.copy(
-            lastModifiedOn = Instant.now().toString(),
-            reviewStatus = ReviewStatus.PENDING
-        )
-        sourceRepository.update(modifiedSource, problem.sourceId)
-        val sourceProblems = problemRepository.getBySourceId(problem.sourceId)
-        sourceProblems.forEach {
-            problemRepository.update(it.copy(skfCode = "", isApproved = false), it.id)
-            if (it.isApproved) {
-                metaService.removeSkfCodeFromUsedList(it.skfCode)
-            }
-        }
+        unapproveSource(source)
     }
 
     override fun updateSource(sourceId: String, sourceData: SourceSubmitDto): SourceDisplayDto {
@@ -240,17 +238,8 @@ class ApprovalServiceImpl: ApprovalService {
             lastModifiedOn = Instant.now().toString(),
             reviewStatus = ReviewStatus.PENDING
         )
-        sourceRepository.update(updatedSource, sourceId)
+        unapproveSource(updatedSource)
         return updatedSource.toDisplayDto(userService.getUsernameById(updatedSource.authorId))
-    }
-
-    override fun updateProblem(
-        problemId: String,
-        problem: ProblemSubmitDto,
-        problemImageFile: MultipartFile?,
-        answerImageFile: MultipartFile?
-    ): Problem {
-        TODO("Not yet implemented")
     }
 
     override fun getSources(): List<SourceDisplayDto> {
@@ -261,5 +250,111 @@ class ApprovalServiceImpl: ApprovalService {
             .map {
                 it.toDisplayDto(userService.getUsernameById(it.authorId))
             }
+    }
+
+    override fun updateProblemTexts(problemId: String, problemTextsDto: ProblemTextsDto) {
+        val userId = userService.getCurrentUserId()
+        val problem = problemRepository.findById(problemId) ?: throw NotFoundException("Problem not found")
+        val source = sourceRepository.findById(problem.sourceId) ?: throw NotFoundException("Source not found")
+        if (source.authorId != userId) {
+            throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
+        }
+        val updatedProblem = problem.copy(
+            sourceListNr = problem.sourceListNr,
+            problemText = problemTextsDto.problemText,
+            answerText = problemTextsDto.answerText,
+        )
+        problemRepository.update(updatedProblem, problemId)
+
+        unapproveSource(source)
+    }
+
+    override fun deleteProblemImage(problemId: String) {
+        val userId = userService.getCurrentUserId()
+        val problem = problemRepository.findById(problemId) ?: throw NotFoundException("Problem not found")
+        val source = sourceRepository.findById(problem.sourceId) ?: throw NotFoundException("Source not found")
+        if (source.authorId != userId) {
+            throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
+        }
+        if (problem.problemImagePath.startsWith("problems/")) {
+            storageRepository.deleteImage(problem.problemImagePath)
+        } else {
+            throw IllegalAccessException("Problem $problemId does not have a problem image")
+        }
+        val updatedProblem = problem.copy(
+            problemImagePath = ""
+        )
+        problemRepository.update(updatedProblem, problemId)
+        unapproveSource(source)
+    }
+
+    override fun deleteAnswerImage(problemId: String) {
+        val userId = userService.getCurrentUserId()
+        val problem = problemRepository.findById(problemId) ?: throw NotFoundException("Problem not found")
+        val source = sourceRepository.findById(problem.sourceId) ?: throw NotFoundException("Source not found")
+        if (source.authorId != userId) {
+            throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
+        }
+        if (problem.answerImagePath.startsWith("answers/")) {
+            storageRepository.deleteImage(problem.answerImagePath)
+        } else {
+            throw IllegalAccessException("Problem $problemId does not have an answer image")
+        }
+        val updatedProblem = problem.copy(
+            answerImagePath = ""
+        )
+        problemRepository.update(updatedProblem, problemId)
+        unapproveSource(source)
+    }
+
+    override fun uploadProblemImage(problemId: String, image: MultipartFile): String {
+        val userId = userService.getCurrentUserId()
+        val problem = problemRepository.findById(problemId) ?: throw NotFoundException("Problem not found")
+        val source = sourceRepository.findById(problem.sourceId) ?: throw NotFoundException("Source not found")
+        if (source.authorId != userId) {
+            throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
+        }
+        val imagesUUID = UUID.randomUUID()
+        val problemImagePath = "problems/${imagesUUID}_${image.originalFilename}"
+        storageRepository.uploadImage(image, problemImagePath)
+        val updatedProblem = problem.copy(
+            problemImagePath = problemImagePath
+        )
+        problemRepository.update(updatedProblem, problemId)
+        unapproveSource(source)
+        return problemImagePath
+    }
+
+    override fun uploadAnswerImage(problemId: String, image: MultipartFile): String {
+        val userId = userService.getCurrentUserId()
+        val problem = problemRepository.findById(problemId) ?: throw NotFoundException("Problem not found")
+        val source = sourceRepository.findById(problem.sourceId) ?: throw NotFoundException("Source not found")
+        if (source.authorId != userId) {
+            throw IllegalAccessException("User $userId does not own source ${problem.sourceId}")
+        }
+        val imagesUUID = UUID.randomUUID()
+        val answerImagePath = "answers/${imagesUUID}_${image.originalFilename}"
+        storageRepository.uploadImage(image, answerImagePath)
+        val updatedProblem = problem.copy(
+            answerImagePath = answerImagePath
+        )
+        problemRepository.update(updatedProblem, problemId)
+        unapproveSource(source)
+        return answerImagePath
+    }
+
+    private fun unapproveSource(source: Source) {
+        val modifiedSource = source.copy(
+            lastModifiedOn = Instant.now().toString(),
+            reviewStatus = ReviewStatus.PENDING
+        )
+        sourceRepository.update(modifiedSource, source.id)
+        val problems = problemRepository.getBySourceId(source.id)
+        problems.forEach {
+            problemRepository.update(it.copy(skfCode = "", isApproved = false), it.id)
+            if (it.isApproved) {
+                metaService.removeSkfCodeFromUsedList(it.skfCode)
+            }
+        }
     }
 }
